test(l1): move FCUv4 targetGasLimit tests to integration suite
Move the payload-id / calc_gas_limit unit tests to
test/tests/blockchain/payload_tests.rs and the PayloadAttributesV4
serde tests to test/tests/rpc/fork_choice_tests.rs (all public API).

Rewrite the validate_attributes_v4 unit tests as end-to-end
ForkChoiceUpdatedV4 handler tests driven through the public RPC entry
point, so the private validator is no longer exposed for testing:
accepts a present targetGasLimit, accepts an absent one (fallback), and
rejects pre-Amsterdam timestamps.
