+++ test/tests/blockchain/mod.rs
@@ -4,4 +4,5 @@ mod eip7702_zero_transfer_tests;
 mod l1_tx_type_tests;
 mod logs_bloom_tests;
 mod mempool_tests;
+mod payload_tests;
 mod smoke_tests;
