+++ test/tests/rpc/fork_choice_tests.rs
@@ -8,12 +8,13 @@ use ethrex_blockchain::{
 };
 use ethrex_common::{
     H160, H256,
-    types::{Block, BlockHeader, DEFAULT_BUILDER_GAS_CEIL, ELASTICITY_MULTIPLIER},
+    types::{Block, BlockHeader, DEFAULT_BUILDER_GAS_CEIL, ELASTICITY_MULTIPLIER, Genesis},
 };
-use ethrex_rpc::engine::fork_choice::ForkChoiceUpdatedV3;
+use ethrex_rpc::engine::fork_choice::{ForkChoiceUpdatedV3, ForkChoiceUpdatedV4};
 use ethrex_rpc::rpc::RpcHandler;
 use ethrex_rpc::test_utils::default_context_with_storage;
-use ethrex_rpc::utils::RpcRequest;
+use ethrex_rpc::types::fork_choice::PayloadAttributesV4;
+use ethrex_rpc::utils::{RpcErr, RpcRequest};
 use ethrex_storage::{EngineType, Store};
 
 fn workspace_root() -> PathBuf {
@@ -123,5 +124,163 @@ async fn test_fcu_v3_finalized_ancestor_returns_valid_with_null_payload_id() {
         response["payloadId"].is_null(),
         "payloadId must be null when head is a finalized ancestor; got {:?}",
         response["payloadId"]
+    );
+}
+
+// execution-apis#796: PayloadAttributesV4 carries a CL-supplied targetGasLimit.
+// Deserialization is permissive (absent / null parse to None); the spec
+// contract is enforced in validate_attributes_v4, exercised by the e2e tests
+// below.
+#[test]
+fn payload_attributes_v4_parses_target_gas_limit_when_present() {
+    let json = serde_json::json!({
+        "timestamp": "0x65",
+        "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000001",
+        "suggestedFeeRecipient": "0x0000000000000000000000000000000000000002",
+        "withdrawals": [],
+        "parentBeaconBlockRoot": "0x0000000000000000000000000000000000000000000000000000000000000003",
+        "slotNumber": "0x10",
+        "targetGasLimit": "0x2faf080",
+    });
+    let attrs: PayloadAttributesV4 = serde_json::from_value(json).unwrap();
+    assert_eq!(attrs.target_gas_limit, Some(50_000_000));
+    assert_eq!(attrs.slot_number, 0x10);
+}
+
+#[test]
+fn payload_attributes_v4_accepts_missing_target_gas_limit() {
+    let json = serde_json::json!({
+        "timestamp": "0x65",
+        "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000001",
+        "suggestedFeeRecipient": "0x0000000000000000000000000000000000000002",
+        "withdrawals": [],
+        "parentBeaconBlockRoot": "0x0000000000000000000000000000000000000000000000000000000000000003",
+        "slotNumber": "0x10",
+    });
+    let attrs: PayloadAttributesV4 = serde_json::from_value(json).unwrap();
+    assert!(attrs.target_gas_limit.is_none());
+}
+
+#[test]
+fn payload_attributes_v4_parses_explicit_null_target_gas_limit() {
+    let json = serde_json::json!({
+        "timestamp": "0x65",
+        "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000001",
+        "suggestedFeeRecipient": "0x0000000000000000000000000000000000000002",
+        "withdrawals": [],
+        "parentBeaconBlockRoot": "0x0000000000000000000000000000000000000000000000000000000000000003",
+        "slotNumber": "0x10",
+        "targetGasLimit": null,
+    });
+    let attrs: PayloadAttributesV4 = serde_json::from_value(json).unwrap();
+    assert!(attrs.target_gas_limit.is_none());
+}
+
+// Builds an in-memory store from l1.json with Amsterdam (= upstream
+// "Glamsterdam") activated at t=0 so the V4 validator paths added by
+// execution-apis#796 are reachable.
+async fn amsterdam_test_store() -> Store {
+    let file = File::open(workspace_root().join("fixtures/genesis/l1.json"))
+        .expect("Failed to open genesis file");
+    let reader = BufReader::new(file);
+    let mut genesis: Genesis =
+        serde_json::from_reader(reader).expect("Failed to deserialize genesis file");
+    genesis.config.amsterdam_time = Some(0);
+    let mut store = Store::new("amsterdam-store.db", EngineType::InMemory)
+        .expect("Failed to build DB for testing");
+    store
+        .add_initial_state(genesis)
+        .await
+        .expect("Failed to add genesis state");
+    store
+}
+
+fn fcu_v4_request(head: H256, timestamp: u64, target_gas_limit: Option<&str>) -> RpcRequest {
+    let target_field = match target_gas_limit {
+        Some(hex) => format!(",\n                    \"targetGasLimit\": \"{hex}\""),
+        None => String::new(),
+    };
+    let body = format!(
+        r#"{{
+            "jsonrpc": "2.0",
+            "method": "engine_forkchoiceUpdatedV4",
+            "params": [
+                {{
+                    "headBlockHash": "{head:#x}",
+                    "safeBlockHash": "{head:#x}",
+                    "finalizedBlockHash": "{head:#x}"
+                }},
+                {{
+                    "timestamp": "{timestamp:#x}",
+                    "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000001",
+                    "suggestedFeeRecipient": "0x0000000000000000000000000000000000000000",
+                    "withdrawals": [],
+                    "parentBeaconBlockRoot": "0x0000000000000000000000000000000000000000000000000000000000000002",
+                    "slotNumber": "0x1"{target_field}
+                }}
+            ],
+            "id": 1
+        }}"#
+    );
+    serde_json::from_str(&body).expect("valid FCU request")
+}
+
+// execution-apis#796: a CL-supplied targetGasLimit on an Amsterdam chain is
+// accepted and the client begins a payload build.
+#[tokio::test]
+async fn fcu_v4_accepts_target_gas_limit_present() {
+    let store = amsterdam_test_store().await;
+    let genesis = store.get_block_header(0).unwrap().unwrap();
+    let request = fcu_v4_request(genesis.hash(), genesis.timestamp + 12, Some("0x2faf080"));
+
+    let context = default_context_with_storage(store).await;
+    let response = ForkChoiceUpdatedV4::call(&request, context)
+        .await
+        .expect("FCU V4 call should succeed");
+
+    assert!(
+        !response["payloadId"].is_null(),
+        "payloadId must be set when V4 attributes are valid; got {:?}",
+        response["payloadId"]
+    );
+}
+
+// We keep targetGasLimit optional for pre-#796 CLs still running on Amsterdam
+// devnets: an absent field is accepted and build_payload_v4 falls back to
+// --builder.gas-limit instead of rejecting.
+#[tokio::test]
+async fn fcu_v4_accepts_target_gas_limit_absent() {
+    let store = amsterdam_test_store().await;
+    let genesis = store.get_block_header(0).unwrap().unwrap();
+    let request = fcu_v4_request(genesis.hash(), genesis.timestamp + 12, None);
+
+    let context = default_context_with_storage(store).await;
+    let response = ForkChoiceUpdatedV4::call(&request, context)
+        .await
+        .expect("FCU V4 call should succeed with fallback gas limit");
+
+    assert!(
+        !response["payloadId"].is_null(),
+        "payloadId must be set when targetGasLimit is absent (fallback); got {:?}",
+        response["payloadId"]
+    );
+}
+
+// V4 attributes for a pre-Amsterdam timestamp are still rejected outright.
+#[tokio::test]
+async fn fcu_v4_rejects_pre_amsterdam_timestamp() {
+    // execution-api.json has no amsterdamTime, so the chain is pre-Amsterdam.
+    let store = test_store().await;
+    let genesis = store.get_block_header(0).unwrap().unwrap();
+    let request = fcu_v4_request(genesis.hash(), genesis.timestamp + 12, Some("0x2faf080"));
+
+    let context = default_context_with_storage(store).await;
+    let err = ForkChoiceUpdatedV4::call(&request, context)
+        .await
+        .expect_err("FCU V4 must reject pre-Amsterdam attributes");
+
+    assert!(
+        matches!(err, RpcErr::InvalidPayloadAttributes(_)),
+        "got: {err:?}"
     );
 }
